pyproject.toml
[project]
name = "home-threat-detection"
version = "0.1.0"
description = "Agentic home threat detection using Google ADK"
authors = [
    {name = "Your Name", email = "your@email.com"},
]
dependencies = [
    "google-adk>=1.15.1",
    "opencv-python>=4.8.0",
    "pillow>=10.1.0",
    "pydantic>=2.5.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "opentelemetry-exporter-gcp-trace>=1.9.0",
    "google-cloud-logging>=3.11.0",
]

requires-python = ">=3.11,<3.14"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = "."